refactor(example): modify example
