+++ example/app/_layout.tsx
@@ -1,5 +1,4 @@
 import { Stack } from "expo-router"
-import { View } from "react-native"
 
 export default function RootLayout() {
 	return <Stack />
